requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudinary==1.29.0
colorama==0.4.4
cryptography==37.0.0
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
pycparser==2.21
pycryptodomex==3.14.1
PyJWT==2.4.0
pymongo==4.1.1
pyotp==2.6.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sib-api-v3-sdk==7.4.0
six==1.16.0
twilio==7.9.0
urllib3==1.26.9
Werkzeug==2.1.2
